refactor: PlaceDetail 패키지 PlaceMap 하위 패키지로 이동
