+++ composeApp/src/commonMain/kotlin/com/daedan/festabook/presentation/placeMap/intent/handler/SelectEventHandlerImpl.kt
@@ -3,7 +3,7 @@ package com.daedan.festabook.presentation.placeMap.intent.handler
 import com.daedan.festabook.di.placeMapHandler.PlaceMapViewModelScope
 import com.daedan.festabook.domain.model.TimeTag
 import com.daedan.festabook.domain.repository.PlaceDetailRepository
-import com.daedan.festabook.presentation.placeDetail.model.toUiModel
+import com.daedan.festabook.presentation.placeMap.placeDetail.model.toUiModel
 import com.daedan.festabook.presentation.placeMap.intent.event.SelectEvent
 import com.daedan.festabook.presentation.placeMap.intent.sideEffect.MapControlSideEffect
 import com.daedan.festabook.presentation.placeMap.intent.sideEffect.PlaceMapSideEffect
