+++ composeApp/src/commonMain/kotlin/com/daedan/festabook/presentation/placeMap/intent/sideEffect/MapControlSideEffect.kt
@@ -1,7 +1,7 @@
 package com.daedan.festabook.presentation.placeMap.intent.sideEffect
 
 import com.daedan.festabook.domain.model.TimeTag
-import com.daedan.festabook.presentation.placeDetail.model.PlaceDetailUiModel
+import com.daedan.festabook.presentation.placeMap.placeDetail.model.PlaceDetailUiModel
 import com.daedan.festabook.presentation.placeMap.intent.state.LoadState
 import com.daedan.festabook.presentation.placeMap.model.InitialMapSettingUiModel
 import com.daedan.festabook.presentation.placeMap.model.PlaceCategoryUiModel
