+++ composeApp/src/commonMain/kotlin/com/daedan/festabook/presentation/placeMap/intent/state/PlaceMapUiState.kt
@@ -1,7 +1,7 @@
 package com.daedan.festabook.presentation.placeMap.intent.state
 
 import com.daedan.festabook.domain.model.TimeTag
-import com.daedan.festabook.presentation.placeDetail.model.PlaceDetailUiModel
+import com.daedan.festabook.presentation.placeMap.placeDetail.model.PlaceDetailUiModel
 import com.daedan.festabook.presentation.placeMap.model.InitialMapSettingUiModel
 import com.daedan.festabook.presentation.placeMap.model.PlaceCategoryUiModel
 import com.daedan.festabook.presentation.placeMap.model.PlaceCoordinateUiModel
