+++ composeApp/src/commonMain/kotlin/com/daedan/festabook/presentation/placeMap/navigation/PlaceMapNavigation.kt
@@ -17,9 +17,9 @@ import androidx.savedstate.read
 import androidx.savedstate.write
 import com.daedan.festabook.presentation.main.FestabookRoute
 import com.daedan.festabook.presentation.main.MainTabRoute
-import com.daedan.festabook.presentation.placeDetail.PlaceDetailViewModel
-import com.daedan.festabook.presentation.placeDetail.component.PlaceDetailRoute
-import com.daedan.festabook.presentation.placeDetail.model.PlaceDetailUiModel
+import com.daedan.festabook.presentation.placeMap.placeDetail.PlaceDetailViewModel
+import com.daedan.festabook.presentation.placeMap.placeDetail.component.PlaceDetailRoute
+import com.daedan.festabook.presentation.placeMap.placeDetail.model.PlaceDetailUiModel
 import com.daedan.festabook.presentation.placeMap.model.PlaceUiModel
 import dev.zacsweers.metrox.viewmodel.assistedMetroViewModel
 import kotlinx.serialization.json.Json
