+++ composeApp/src/commonMain/kotlin/com/daedan/festabook/presentation/placeMap/placeDetail/component/PlaceDetailScreen.kt
@@ -1,4 +1,4 @@
-package com.daedan.festabook.presentation.placeDetail.component
+package com.daedan.festabook.presentation.placeMap.placeDetail.component
 
 import androidx.compose.animation.animateContentSize
 import androidx.compose.animation.core.Spring
@@ -52,10 +52,10 @@ import com.daedan.festabook.presentation.common.component.EmptyStateScreen
 import com.daedan.festabook.presentation.common.component.FestabookImage
 import com.daedan.festabook.presentation.common.component.LoadingStateScreen
 import com.daedan.festabook.presentation.common.component.URLText
-import com.daedan.festabook.presentation.placeDetail.PlaceDetailViewModel
-import com.daedan.festabook.presentation.placeDetail.model.ImageUiModel
-import com.daedan.festabook.presentation.placeDetail.model.PlaceDetailUiModel
-import com.daedan.festabook.presentation.placeDetail.model.PlaceDetailUiState
+import com.daedan.festabook.presentation.placeMap.placeDetail.PlaceDetailViewModel
+import com.daedan.festabook.presentation.placeMap.placeDetail.model.ImageUiModel
+import com.daedan.festabook.presentation.placeMap.placeDetail.model.PlaceDetailUiModel
+import com.daedan.festabook.presentation.placeMap.placeDetail.model.PlaceDetailUiState
 import com.daedan.festabook.presentation.placeMap.component.PlaceCategoryLabel
 import com.daedan.festabook.presentation.placeMap.model.PlaceCategoryUiModel
 import com.daedan.festabook.presentation.placeMap.model.PlaceUiModel
