+++ composeApp/src/commonTest/kotlin/com/daedan/festabook/viewModel/placeMap/placeDetail/PlaceDetailViewModelTest.kt
@@ -1,12 +1,12 @@
-package com.daedan.festabook.placeDetail
+package com.daedan.festabook.placeMap.placeDetail
 
 import com.daedan.festabook.domain.repository.PlaceDetailRepository
 import com.daedan.festabook.news.FAKE_NOTICES
 import com.daedan.festabook.placeMap.FAKE_PLACES
 import com.daedan.festabook.presentation.news.notice.model.toUiModel
-import com.daedan.festabook.presentation.placeDetail.PlaceDetailViewModel
-import com.daedan.festabook.presentation.placeDetail.model.PlaceDetailUiState
-import com.daedan.festabook.presentation.placeDetail.model.toUiModel
+import com.daedan.festabook.presentation.placeMap.placeDetail.PlaceDetailViewModel
+import com.daedan.festabook.presentation.placeMap.placeDetail.model.PlaceDetailUiState
+import com.daedan.festabook.presentation.placeMap.placeDetail.model.toUiModel
 import com.daedan.festabook.presentation.placeMap.model.toUiModel
 import dev.mokkery.answering.returns
 import dev.mokkery.everySuspend
